ddrs_api/requirements.txt
asgiref==3.9.1
cffi==1.17.1
cryptography==45.0.5
Django==5.2.4
django-filter==25.1
djangorestframework==3.16.0
psycopg2-binary==2.9.10
pycparser==2.22
python-decouple==3.8
sqlparse==0.5.3
